works now

+++ datasets/dynaMO/dataset.py
@@ -131,7 +131,7 @@ class ToTensor(object):
         # torch image: C X H X W
         image = image.transpose((2, 0, 1))
         return {'image': torch.from_numpy(image).type(torch.float32),
-                'target': torch.from_numpy(target).type(torch.float32)}
+                'target': torch.from_numpy(target).type(torch.int64)}
